refactor: change the direction path
