+++ pages/index.tsx
@@ -1,4 +1,4 @@
-import { HomeScreen } from "@/Home";
+import { HomeScreen } from "@/views/Home";
 
 export default function HomePage() {
   return <HomeScreen />;
